also promote ndatum in MyPoolChunk class

+++ src/my_pool_chunk.h
@@ -18,8 +18,8 @@ namespace LAMMPS_NS {
 
 template <class T> class MyPoolChunk {
  public:
-  int ndatum;    // total # of stored datums
-  int nchunk;    // total # of stored chunks
+  bigint ndatum;    // total # of stored datums
+  int nchunk;       // total # of stored chunks
 
   MyPoolChunk(int user_minchunk = 1, int user_maxchunk = 1, int user_nbin = 1,
               int user_chunkperpage = 1024, int user_pagedelta = 1);
